beautified and added bug tracker to package.json

+++ commands/search.js
@@ -18,7 +18,7 @@ exports.run = async (client, message, args) => {
       .reply("You need to join a voice channel first!")
       .catch(client.logger.error);
 
-  const search = args.join(" "); 
+  const search = args.join(" ");
   const resultsEmbed = new MessageEmbed()
     .setTitle(`**Reply with the song number you want to play**`)
     .setDescription(`Results for: ${search}`)
@@ -50,8 +50,7 @@ exports.run = async (client, message, args) => {
 
     message.channel.activeCollector = false;
     client.commands.get("play").run(client, message, [choice]);
-    if (resultsMessage)
-      resultsMessage.delete().catch(client.logger.error);
+    if (resultsMessage) resultsMessage.delete().catch(client.logger.error);
   } catch (error) {
     client.logger.error(error);
     message.channel.activeCollector = false;
